pyproject.toml
[project]
name = "bizantine-watcher"
version = "0.1.0"
description = ""
authors = [
    { name = "gabrieldiem", email = "gabriel.diem5@gmail.com" },
]
requires-python = ">=3.13,<4.0"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
openai = "^2.30.0"
pyyaml = "^6.0.3"
python-dotenv = "^1.2.2"
fastapi = "^0.135.2"
uvicorn = "^0.42.0"
jinja2 = "^3.1.6"
httpx = ">=0.27"

[tool.poetry.group.dev.dependencies]
pylint = "^4.0.5"
mypy = "^1.19.1"
black = "^26.3.1"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
